1. feat: Complete the UI design of profile screen. 2. feat: Add shimmer effect on home and profile screen for loading state. 3. feat: Create extensions and common common screen files.

+++ app/src/main/java/com/github/almasud/rick_and_morty/ui/nav_graph/home_nav_graph.kt
@@ -63,7 +63,7 @@ fun HomeNavGraph(navController: NavHostController) {
                 })
         ) { navBackStackEntry ->
             navBackStackEntry.arguments?.getString(App.Constant.Navigation.Argument.CHARACTER_ID)?.let {
-                ProfileScreen(characterId = it)
+                ProfileScreen(characterId = it.toInt())
             }
         }
     }
